requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
cssselect==1.3.0
cssutils==2.11.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
environs==14.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
lxml==5.4.0
marshmallow==3.26.1
more-itertools==10.7.0
packaging==24.2
premailer==3.10.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.9.0
